requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
httpx==0.23.3
django-environ==0.9.0